Agentic-RAG-LangChain-main/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
cachetools==5.4.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.2.2
cryptography==42.0.8
dataclasses-json==0.6.7
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
Django==5.1.3
djangorestframework==3.15.2
Everything-Tkinter==0.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
flatbuffers==25.1.21
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.3.0
gast==0.6.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
greenlet==3.1.1
groq==0.20.0
grpcio==1.70.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
idna==3.4
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.32.0
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.9.0
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==2.15.0
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.48
langchain-groq==0.3.1
langchain-huggingface==0.1.2
langchain-mistralai==0.2.9
langchain-openai==0.3.10
langchain-pinecone==0.2.3
langchain-tests==0.3.16
langchain-text-splitters==0.3.7
langchainhub==0.1.21
langsmith==0.3.18
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.1
ml-dtypes==0.2.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.2.0
mypy-extensions==1.0.0
mysqlclient==2.2.5
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
openai==1.68.2
opencv-python-headless==4.7.0.72
opt_einsum==3.4.0
optree==0.14.0
orjson==3.10.16
overrides==7.7.0
packaging==24.2
paho-mqtt==1.6.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
phonenumbers==8.13.35
pillow==11.1.0
pinecone==5.4.2
pinecone-client==6.0.0
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==4.25.6
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.13
PyAutoGUI==0.9.53
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyGetWindow==0.0.9
Pygments==2.19.1
PyJWT==2.8.0
PyMsgBox==1.0.9
PyMuPDF==1.25.3
pyparsing==3.1.2
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
pytweening==1.0.4
pytz==2024.2
pywin32==308
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.1
Send2Trash==1.8.3
sentence-transformers==3.4.1
six==1.17.0
sniffio==1.3.1
sockets==1.0.0
soupsieve==2.6
SQLAlchemy==2.0.39
sqlparse==0.5.1
stack-data==0.6.3
sympy==1.13.1
syrupy==4.9.1
tenacity==9.0.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_keras==0.1
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.0
types-python-dateutil==2.9.0.20241206
types-requests==2.32.0.20250306
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.3.0
vidstream==0.0.14
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.1.3
wrapt==1.14.1
yarl==1.18.3
zstandard==0.23.0
streamlit